+++ framework/Furion/UnifyResult/Providers/IUnifyResultProvider.cs
@@ -42,7 +42,8 @@ namespace Furion.UnifyResult
         /// </summary>
         /// <param name="context"></param>
         /// <param name="statusCode"></param>
+        /// <param name="options"></param>
         /// <returns></returns>
-        Task OnResponseStatusCodes(HttpContext context, int statusCode);
+        Task OnResponseStatusCodes(HttpContext context, int statusCode, UnifyResultStatusCodesOptions options = default);
     }
 }